[certs] Rework SSLHostConfig so tomcat 7/8 startup
