+++ core/src/main/java/psiprobe/controllers/certificates/ListCertificatesController.java
@@ -11,11 +11,10 @@
 package psiprobe.controllers.certificates;
 
 import org.apache.catalina.connector.Connector;
-import org.apache.commons.beanutils.BeanUtils;
+import org.apache.commons.beanutils.MethodUtils;
 import org.apache.coyote.ProtocolHandler;
 import org.apache.coyote.http11.AbstractHttp11JsseProtocol;
-import org.apache.tomcat.util.net.SSLHostConfig;
-import org.apache.tomcat.util.net.SSLHostConfigCertificate;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -42,8 +41,6 @@ import java.security.cert.X509Certificate;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Set;
-
 import javax.management.ObjectName;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -225,55 +222,23 @@ public class ListCertificatesController extends AbstractTomcatContainerControlle
       throws IllegalAccessException, InvocationTargetException {
     ConnectorInfo info = new ConnectorInfo();
     info.setName(ObjectName.unquote(protocol.getName()));
-    info.setDefaultSSLHostConfigName(protocol.getDefaultSSLHostConfigName());
 
-    SSLHostConfig[] sslHostConfigs = protocol.findSslHostConfigs();
-    List<SSLHostConfigInfo> sslHostConfigInfos = new ArrayList<>(sslHostConfigs.length);
-    info.setSslHostConfigInfos(sslHostConfigInfos);
-
-    for (SSLHostConfig sslHostConfig : sslHostConfigs) {
-      sslHostConfigInfos.add(toSSLHostConfigInfo(sslHostConfig));
+    try {
+      // Introduced in Tomcat 8.5.x+
+      Object defaultSSLHostConfigName = MethodUtils.invokeMethod(protocol, "getDefaultSSLHostConfigName", null);
+      if (defaultSSLHostConfigName == null) {
+          // We are using Tomcat 7, fill in the old way
+          // TODO Build this out...
+          return info;
+      }
+      info.setDefaultSSLHostConfigName(String.valueOf(defaultSSLHostConfigName));
+      new SSLHostConfigHelper(protocol, info);
+    } catch (NoSuchMethodException e) {
+        logger.error("Cannot determine defaultSSLHostConfigName");
+        logger.trace("", e);
     }
 
     return info;
-  }
-
-  /**
-   * To SSLHostConfig info.
-   * 
-   * @param sslHostConfig the SSLHostConfig
-   * @return the SSLHostConfig info
-   * @throws IllegalAccessException the illegal access exception
-   * @throws InvocationTargetException the invocation target exception
-   */
-  private SSLHostConfigInfo toSSLHostConfigInfo(SSLHostConfig sslHostConfig)
-      throws IllegalAccessException, InvocationTargetException {
-    SSLHostConfigInfo sslHostConfigInfo = new SSLHostConfigInfo();
-    BeanUtils.copyProperties(sslHostConfigInfo, sslHostConfig);
-
-    Set<SSLHostConfigCertificate> certificates = sslHostConfig.getCertificates();
-    List<CertificateInfo> certificateInfos = new ArrayList<>(certificates.size());
-    sslHostConfigInfo.setCertificateInfos(certificateInfos);
-    for (SSLHostConfigCertificate sslHostConfigCertificate : certificates) {
-      certificateInfos.add(toCertificateInfo(sslHostConfigCertificate));
-    }
-
-    return sslHostConfigInfo;
-  }
-
-  /**
-   * To certificate info.
-   * 
-   * @param sslHostConfigCertificate the SSLHostConfigCertificate
-   * @return the certificate info
-   * @throws IllegalAccessException the illegal access exception
-   * @throws InvocationTargetException the invocation target exception
-   */
-  private CertificateInfo toCertificateInfo(SSLHostConfigCertificate sslHostConfigCertificate)
-      throws IllegalAccessException, InvocationTargetException {
-    CertificateInfo certificateInfo = new CertificateInfo();
-    BeanUtils.copyProperties(certificateInfo, sslHostConfigCertificate);
-    return certificateInfo;
   }
 
   /**
